requirements.txt
# Web Scraping Core
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
html5lib==1.1

# Data Processing & Storage
pandas==2.1.4
openpyxl==3.1.2
sqlite3  # Built into Python

# HTTP & Session Management
requests-session==0.1.2
urllib3==2.1.0

# Concurrency & Performance
asyncio  # Built into Python
threading  # Built into Python
concurrent.futures  # Built into Python

# Configuration & Logging
pyyaml==6.0.1
python-dotenv==1.0.0
colorlog==6.8.0

# Data Validation & Parsing
jsonschema==4.20.0
python-dateutil==2.8.2
validators==0.22.0

# Optional: Advanced Features
selenium==4.15.2  # For JavaScript-heavy sites
fake-useragent==1.4.0  # Random user agents
retry==0.9.2  # Retry mechanisms

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1